Move changes to the preview style
