+++ tests/data/miscellaneous/stub.pyi
@@ -56,13 +56,6 @@ class Nested:
         def who_has_to_know(self): ...
     def verse(self): ...
 
-class Outer:
-    class InnerStub: ...
-    outer_attr_after_inner_stub: int
-    class Inner:
-        inner_attr: int
-    outer_attr: int
-
 class Conditional:
     def f(self): ...
     if sys.version_info >= (3, 8):
@@ -135,15 +128,6 @@ class Nested:
         def who_has_to_know(self): ...
 
     def verse(self): ...
-
-class Outer:
-    class InnerStub: ...
-    outer_attr_after_inner_stub: int
-
-    class Inner:
-        inner_attr: int
-
-    outer_attr: int
 
 class Conditional:
     def f(self): ...
